+++ VIStk/Widgets/__init__.py
@@ -8,6 +8,11 @@ from VIStk.Widgets._TabBar import TabBar
 from VIStk.Widgets._HostMenu import HostMenu
 from VIStk.Widgets._InfoRow import InfoRow
 from VIStk.Widgets._SplitView import SplitView
+from VIStk.Widgets._Dialogs import confirm, confirm_discard
+from VIStk.Widgets._Tooltip import Tooltip
+from VIStk.Widgets._CollapsibleFrame import CollapsibleFrame
+from VIStk.Widgets._AutocompleteEntry import AutocompleteEntry
+from VIStk.Widgets._DateEntry import DateEntry
 
 
 __all__ = ["VISMenu",
@@ -20,4 +25,10 @@ __all__ = ["VISMenu",
            "HostMenu",
            "InfoRow",
            "SplitView",
+           "confirm",
+           "confirm_discard",
+           "Tooltip",
+           "CollapsibleFrame",
+           "AutocompleteEntry",
+           "DateEntry",
            ]